+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Please also have a look at our
 
 ### Fixed
 
+- Clear all previous caches on each invokation to avoid memory usage build-up
+  (#1567)
 - Allow very deeply-nested HTML (#1556, #1560)
 
 ## 8.2.0: Add support for Symfony 8.0