+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Please also have a look at our
 
 ### Changed
 
+- Require `sabberworm/php-css-parser` >= 9.2.0 (#1575)
+
 ### Deprecated
 
 ### Removed